refactor: replace ApiError class with enum
