+++ server/src/types/api-error.enum.ts
@@ -1,0 +1,12 @@
+export enum ApiError {
+  UNEXPECTED_ERROR = "UNEXPECTED_ERROR",
+
+  // Authentication errors
+  NO_SESSION_COOKIE = "NO_SESSION_COOKIE",
+  TOKEN_EXPIRED = "TOKEN_EXPIRED",
+  INVALID_TOKEN = "INVALID_TOKEN",
+
+  // Validation errors
+  INVALID_USERNAME = "INVALID_USERNAME",
+  INVALID_PASSWORD = "INVALID_PASSWORD",
+}
